+++ packages/frontend/tests/stories/activities.spec.ts
@@ -6,7 +6,7 @@ test.describe('Activity history - transmissions and activities', () => {
     await page.goto(appUrlWithPlaywrightId('activity-history'));
     // Go to details for dialog with activity history
     await page.getByRole('link', { name: 'This has a sender name defined' }).click();
-    await expect(page.getByRole('link', { name: 'Tilbake' })).toBeVisible();
+    await expect(page.getByRole('link', { name: 'Tilbake', exact: true })).toBeVisible();
 
     // Open modal dialog with activity history