+++ CHANGELOG.md
@@ -1,5 +1,118 @@
 # Changelog
 
+## [1.66.2](https://github.com/Altinn/dialogporten-frontend/compare/v1.66.1...v1.66.2) (2025-09-12)
+
+
+### Bug Fixes
+
+* **bff:** add dialogporten to health check ([#2704](https://github.com/Altinn/dialogporten-frontend/issues/2704)) ([f6c11db](https://github.com/Altinn/dialogporten-frontend/commit/f6c11dbe0c8712a8c6eb1889246c0c5323290d73))
+* use sender in filters instead of senders from 100 first dialogs ([#2710](https://github.com/Altinn/dialogporten-frontend/issues/2710)) ([1eb3d79](https://github.com/Altinn/dialogporten-frontend/commit/1eb3d792914c67b06da44d381152448887be014b))
+
+## [1.66.1](https://github.com/Altinn/dialogporten-frontend/compare/v1.66.0...v1.66.1) (2025-09-12)
+
+
+### Bug Fixes
+
+* tour not showing ([#2702](https://github.com/Altinn/dialogporten-frontend/issues/2702)) ([66d4a7f](https://github.com/Altinn/dialogporten-frontend/commit/66d4a7f0d80ef3aac27d671f99bd389235da7c30))
+
+## [1.66.0](https://github.com/Altinn/dialogporten-frontend/compare/v1.65.3...v1.66.0) (2025-09-12)
+
+
+### Features
+
+* New modal design. Improved logic for notification settings. ([daabc56](https://github.com/Altinn/dialogporten-frontend/commit/daabc566dee85e1484eb28a5af7337ecb92134fd))
+
+## [1.65.3](https://github.com/Altinn/dialogporten-frontend/compare/v1.65.2...v1.65.3) (2025-09-12)
+
+
+### Bug Fixes
+
+* **deps:** update dependency typeorm to v0.3.26 ([#2288](https://github.com/Altinn/dialogporten-frontend/issues/2288)) ([c6e2abe](https://github.com/Altinn/dialogporten-frontend/commit/c6e2abe68dfdbc277cf36d552bd0daf688f558a0))
+* prevent cursor jump in search input ([#2697](https://github.com/Altinn/dialogporten-frontend/issues/2697)) ([e51aeb9](https://github.com/Altinn/dialogporten-frontend/commit/e51aeb93d375892975b06f199a88a1c6842eea68))
+
+## [1.65.2](https://github.com/Altinn/dialogporten-frontend/compare/v1.65.1...v1.65.2) (2025-09-11)
+
+
+### Bug Fixes
+
+* add test ids to sidebar menu items ([#2691](https://github.com/Altinn/dialogporten-frontend/issues/2691)) ([1d882f1](https://github.com/Altinn/dialogporten-frontend/commit/1d882f11f2d2264bf81ed0a1b32097bcccfef9bc))
+* **autocomplete:** avoid duplicate params when org matches query ([#2689](https://github.com/Altinn/dialogporten-frontend/issues/2689)) ([a4974c4](https://github.com/Altinn/dialogporten-frontend/commit/a4974c445837024942a90f92bc428562692c45d5))
+
+## [1.65.1](https://github.com/Altinn/dialogporten-frontend/compare/v1.65.0...v1.65.1) (2025-09-09)
+
+
+### Bug Fixes
+
+* wait on subscription to be established before loading fce ([#2671](https://github.com/Altinn/dialogporten-frontend/issues/2671)) ([d2a84e1](https://github.com/Altinn/dialogporten-frontend/commit/d2a84e17c88e0d21e06ae6645e75f438ad2acd17))
+
+## [1.65.0](https://github.com/Altinn/dialogporten-frontend/compare/v1.64.3...v1.65.0) (2025-09-08)
+
+
+### Features
+
+* support for github flavored markdown for table support in markdown/html ([#2669](https://github.com/Altinn/dialogporten-frontend/issues/2669)) ([611594f](https://github.com/Altinn/dialogporten-frontend/commit/611594f781eea4b92d044de90fe4bdddedce997e))
+
+## [1.64.3](https://github.com/Altinn/dialogporten-frontend/compare/v1.64.2...v1.64.3) (2025-09-08)
+
+
+### Bug Fixes
+
+* group search results into a single group with a description ([#2662](https://github.com/Altinn/dialogporten-frontend/issues/2662)) ([ab8d638](https://github.com/Altinn/dialogporten-frontend/commit/ab8d6384381bb21b76f55e253d50336656bfffc9))
+* Update maintenance page support email ([#2660](https://github.com/Altinn/dialogporten-frontend/issues/2660)) ([d1ce046](https://github.com/Altinn/dialogporten-frontend/commit/d1ce04614b404e3ce9b3ed058375b54b4ba6d029))
+
+## [1.64.2](https://github.com/Altinn/dialogporten-frontend/compare/v1.64.1...v1.64.2) (2025-09-05)
+
+
+### Bug Fixes
+
+* set secure cookie only when session exists ([#2656](https://github.com/Altinn/dialogporten-frontend/issues/2656)) ([613cf34](https://github.com/Altinn/dialogporten-frontend/commit/613cf346d406120aed10c1a5f4a75604d820dd88))
+
+## [1.64.1](https://github.com/Altinn/dialogporten-frontend/compare/v1.64.0...v1.64.1) (2025-09-04)
+
+
+### Bug Fixes
+
+* **bff:** ensure security headers are always exposed ([#2651](https://github.com/Altinn/dialogporten-frontend/issues/2651)) ([348b440](https://github.com/Altinn/dialogporten-frontend/commit/348b440d5eb7d1a955f3a0c0cc4975deb24136f0))
+* ensure gql requests are instrumented properly ([#2653](https://github.com/Altinn/dialogporten-frontend/issues/2653)) ([05f9182](https://github.com/Altinn/dialogporten-frontend/commit/05f9182ca0bed08dafa513e006abbdf7342ce1a5))
+* improvements to searchbar ([#2648](https://github.com/Altinn/dialogporten-frontend/issues/2648)) ([272c2c3](https://github.com/Altinn/dialogporten-frontend/commit/272c2c3f9701f3eebe3702f17327e1babc7ea235))
+
+## [1.64.0](https://github.com/Altinn/dialogporten-frontend/compare/v1.63.6...v1.64.0) (2025-09-03)
+
+
+### Features
+
+* **frontend:** add tracking for dialog actions ([#2646](https://github.com/Altinn/dialogporten-frontend/issues/2646)) ([eb9ad76](https://github.com/Altinn/dialogporten-frontend/commit/eb9ad765e16c6025d177aeba15cc61453847e2c9))
+
+
+### Bug Fixes
+
+* **app-insights:** avoid tracing every http request ([#2636](https://github.com/Altinn/dialogporten-frontend/issues/2636)) ([84c677b](https://github.com/Altinn/dialogporten-frontend/commit/84c677b28ea3eacc45aecf2139af3648eb02bdfa))
+* **frontend:** use correct way of setting headers when tracking ([#2647](https://github.com/Altinn/dialogporten-frontend/issues/2647)) ([fb94d13](https://github.com/Altinn/dialogporten-frontend/commit/fb94d13d43981aaef191c633213f7710068fc1c6))
+
+## [1.63.6](https://github.com/Altinn/dialogporten-frontend/compare/v1.63.5...v1.63.6) (2025-09-01)
+
+
+### Bug Fixes
+
+* beta modal navigation to onboarding missing ctx ([#2621](https://github.com/Altinn/dialogporten-frontend/issues/2621)) ([1a65565](https://github.com/Altinn/dialogporten-frontend/commit/1a655659cac64ef2a212f16e0e4b393f7eedb400))
+* do not fetch isAuthenticated in background ([#2623](https://github.com/Altinn/dialogporten-frontend/issues/2623)) ([63019be](https://github.com/Altinn/dialogporten-frontend/commit/63019be2e9c2f5d6516dd968dc2a3677daaca9c2))
+
+## [1.63.5](https://github.com/Altinn/dialogporten-frontend/compare/v1.63.4...v1.63.5) (2025-08-29)
+
+
+### Bug Fixes
+
+* adjustments to texts in folders ([#2617](https://github.com/Altinn/dialogporten-frontend/issues/2617)) ([6b9878e](https://github.com/Altinn/dialogporten-frontend/commit/6b9878eb12f757771106aa56f72577011e314f67))
+
+## [1.63.4](https://github.com/Altinn/dialogporten-frontend/compare/v1.63.3...v1.63.4) (2025-08-29)
+
+
+### Bug Fixes
+
+* add links to about the new Altinn ([#2614](https://github.com/Altinn/dialogporten-frontend/issues/2614)) ([576c970](https://github.com/Altinn/dialogporten-frontend/commit/576c970e1daa869fc6b382825b80e14a0bb981ea))
+* change translations for awaiting and not_applicable ([#2615](https://github.com/Altinn/dialogporten-frontend/issues/2615)) ([cebbadc](https://github.com/Altinn/dialogporten-frontend/commit/cebbadc31defcfcc5cb7af49ba05a8f05444a6aa))
+* update texts for folder and use take me back in stead of leave beta as link text ([#2612](https://github.com/Altinn/dialogporten-frontend/issues/2612)) ([ce3e20b](https://github.com/Altinn/dialogporten-frontend/commit/ce3e20bca2b40132e93c797ce95c384454201808))
+
 ## [1.63.3](https://github.com/Altinn/dialogporten-frontend/compare/v1.63.2...v1.63.3) (2025-08-29)